third_party/blink/web_tests/html5lib/generated/run-tests7-write-expected.txt

CONSOLE ERROR: Uncaught SyntaxError: Failed to execute 'write' on 'Document': Unexpected token '<'
CONSOLE WARNING: A <select> tag was parsed within another <select> tag and was converted into </select>. Please add the missing </select> end tag.
../resources/tests7.dat:
34

Test 34 of 34 in ../resources/tests7.dat failed. Input:
<select><keygen>
Got:
| <html>
|   <head>
|   <body>
|     <select>
|       <keygen>
Expected:
| <html>
|   <head>
|   <body>
|     <select>
|     <keygen>
